remove FindGroup.py
